menambahkan fitur export excel

+++ rak_belakang.php
@@ -69,8 +69,7 @@ function validasi_input($data)
 
 ?>
 
-<div class="modal fade bd-example-modal-lg" tabindex="-1" role="dialog" aria-labelledby="myLargeModalLabel"
-  aria-hidden="true">
+<div class="modal fade bd-example-modal-lg" tabindex="-1" role="dialog" aria-labelledby="myLargeModalLabel" aria-hidden="true">
   <div class="modal-dialog ">
     <div class="modal-content">
       <div class="modal-header">
@@ -105,15 +104,13 @@ function validasi_input($data)
             <div class="col-md-6">
               <div class="d-flex align-items-centers justify-content-center">
                 <label for="meker1">HEWRITCH</label>
-                <input type="radio" name="meker" id="meker1" class=" ml-2" value="HERWITCH"
-                  style="width:25px; height:25px;cursor:pointer" required>
+                <input type="radio" name="meker" id="meker1" class=" ml-2" value="HERWITCH" style="width:25px; height:25px;cursor:pointer" required>
               </div>
             </div>
             <div class="col-md-6 ">
               <div class="d-flex align-items-centers justify-content-center">
                 <label for="meker2">HYCASHT</label>
-                <input type="radio" name="meker" id="meker2" class=" ml-2" value="HYCASHT"
-                  style="width:25px; height:25px;cursor:pointer">
+                <input type="radio" name="meker" id="meker2" class=" ml-2" value="HYCASHT" style="width:25px; height:25px;cursor:pointer">
               </div>
             </div>
           </div>
@@ -135,16 +132,14 @@ function validasi_input($data)
             <div class="col-md-6">
               <div class="d-flex align-items-centers justify-content-center">
                 <label for="meker1">ALCPL</label>
-                <input type="radio" name="pilihan_penempatan" id="meker1" class=" ml-2" value="ALCPL"
-                  style="width:25px; height:25px;cursor:pointer">
+                <input type="radio" name="pilihan_penempatan" id="meker1" class=" ml-2" value="ALCPL" style="width:25px; height:25px;cursor:pointer">
               </div>
 
             </div>
             <div class="col-md-6 ">
               <div class="d-flex align-items-centers justify-content-center">
                 <label for="meker2">BCPL</label>
-                <input type="radio" name="pilihan_penempatan" id="meker2" class=" ml-2" value="BCPL"
-                  style="width:25px; height:25px;cursor:pointer">
+                <input type="radio" name="pilihan_penempatan" id="meker2" class=" ml-2" value="BCPL" style="width:25px; height:25px;cursor:pointer">
               </div>
             </div>
           </div>
@@ -167,14 +162,19 @@ function validasi_input($data)
   </div>
 </div>
 
-
+<div class="card p-2 m-4">
+  <h3>Data Rak Belakang</h3>
+</div>
 <div class="card p-4 m-4">
-  <div style="width:150px" class="mb-4">
+  <div class="mb-4 d-flex justify-content-between">
 
     <button type="button" class="btn btn-primary" data-toggle="modal" data-target=".bd-example-modal-lg"> + Tambah
       Data</button>
     <!-- <a type="button" href="cetak-rakdepan.php" class="btn btn-primary"> + Cetak PDF
     </a> -->
+    <a type="button" target="_blankc" href="export-excel-rakbelakang.php" class="btn btn-primary"> Export Excel
+      <i class="fas fa-file-excel"></i>
+    </a>
   </div>
   <table id="example" class="table table-striped table-bordered" style="width:100%">
     <thead>
@@ -197,27 +197,24 @@ function validasi_input($data)
       $no = 1;
       while ($row = mysqli_fetch_assoc($query)) {
       ?>
-      <tr>
-        <td><?= $no++ ?></td>
-        <td><?= DateTime::createFromFormat('Y-m-d', $row['tanggal'])->format('d-m-Y') ?></td>
-        <td><?= $row['no_resi'] ?></td>
-        <td><?= $row['nama_sparepart'] ?></td>
-        <td><?= $row['meker'] ?></td>
-        <td><?= $row['no_rak'] ?></td>
-        <td><?= $row['jumlah'] ?></td>
-        <td><?= $row['penempatan'] ?></td>
-        <td align="center" width="15%">
-          <a href="index.php?page=edit-rak-belakang&id=<?= $row['id'] ?>" data-toggle="tooltip" title="Edit"
-            style="background:  #2A3042;color:#ffffff;" class="d-none  d-sm-inline-block btn btn-sm shadow-sm">
-            <i class="fas fa-edit fa-sm text-white-50"></i>
-          </a>
-          <a href="delete-rak-belakang.php?id=<?= $row['id'] ?>" data-toggle="tooltip" title="Hapus"
-            onclick="return confirm('Yakin Ingin menghapus data?')" style="background: red"
-            class="d-none d-sm-inline-block btn btn-sm  shadow-sm">
-            <i class="fas fa-trash-alt fa-sm text-white-50"></i>
-          </a>
-        </td>
-      </tr>
+        <tr>
+          <td><?= $no++ ?></td>
+          <td><?= DateTime::createFromFormat('Y-m-d', $row['tanggal'])->format('d-m-Y') ?></td>
+          <td><?= $row['no_resi'] ?></td>
+          <td><?= $row['nama_sparepart'] ?></td>
+          <td><?= $row['meker'] ?></td>
+          <td><?= $row['no_rak'] ?></td>
+          <td><?= $row['jumlah'] ?></td>
+          <td><?= $row['penempatan'] ?></td>
+          <td align="center" width="15%">
+            <a href="index.php?page=edit-rak-belakang&id=<?= $row['id'] ?>" data-toggle="tooltip" title="Edit" style="background:  #2A3042;color:#ffffff;" class="d-none  d-sm-inline-block btn btn-sm shadow-sm">
+              <i class="fas fa-edit fa-sm text-white-50"></i>
+            </a>
+            <a href="delete-rak-belakang.php?id=<?= $row['id'] ?>" data-toggle="tooltip" title="Hapus" onclick="return confirm('Yakin Ingin menghapus data?')" style="background: red" class="d-none d-sm-inline-block btn btn-sm  shadow-sm">
+              <i class="fas fa-trash-alt fa-sm text-white-50"></i>
+            </a>
+          </td>
+        </tr>
       <?php
       }
       ?>
@@ -239,13 +236,13 @@ function validasi_input($data)
 </div>
 
 <script>
-$(document).ready(function() {
-  $('#example').DataTable();
+  $(document).ready(function() {
+    $('#example').DataTable();
 
-  $('input[type=radio][name=pilihan_penempatan]').change(function() {
-    if (this.value == 'ALCPL') {
-      $('#alcpl').html(
-        `
+    $('input[type=radio][name=pilihan_penempatan]').change(function() {
+      if (this.value == 'ALCPL') {
+        $('#alcpl').html(
+          `
         <div id="alcpl-child">
          <div class="d-flex align-items-centers justify-content-center">
               <label for="meker1">ACD</label>
@@ -269,11 +266,11 @@ $(document).ready(function() {
                 style="width:25px; height:25px;cursor:pointer">
             </div>
             </div>`
-      )
-      $('#bcpl-child').remove()
-    } else if (this.value == 'BCPL') {
-      $('#bcpl').html(
-        `<div id="bcpl-child">
+        )
+        $('#bcpl-child').remove()
+      } else if (this.value == 'BCPL') {
+        $('#bcpl').html(
+          `<div id="bcpl-child">
         <div class="d-flex align-items-centers justify-content-center">
               <label for="meker5">CHARGIN MACHINE</label>
               <input type="radio" name="penempatan" id="meker5" class=" ml-2" value="BCPL Charging Machine"
@@ -297,11 +294,11 @@ $(document).ready(function() {
                 style="width:25px; height:25px;cursor:pointer">
             </div>
             </div>`
-      )
-      $('#alcpl-child').remove()
+        )
+        $('#alcpl-child').remove()
 
-    }
+      }
+    });
+
   });
-
-});
 </script>
